[#24] Radius 단위 CGFloat로 변환

+++ Handy/Handy/Source/Atom/HandyButton/HandyBoxButton.swift
@@ -119,7 +119,7 @@ public class HandyBoxButton: UIButton, HandyButtonProtocol {
             }
         }
 
-        fileprivate var rounding: Int {
+        fileprivate var rounding: CGFloat {
             switch self {
             case .xLarge, .large:
                 return HandySemantic.radiusXL
@@ -448,7 +448,7 @@ public class HandyBoxButton: UIButton, HandyButtonProtocol {
      버튼의 라운딩 값을 세팅합니다.
      */
     private func setBoxButtonRounding() {
-        self.layer.cornerRadius = CGFloat(size.rounding)
+        self.layer.cornerRadius = size.rounding
     }
 
     public override func layoutSubviews() {
